move messages, remove worldguard
